+++ packages/nuxt/test/enums.ts
@@ -1,0 +1,6 @@
+export enum TestId {
+  AutoSizes = 'auto-sizes',
+  ModuleSizes = 'module-sizes',
+  PropSizes = 'prop-sizes',
+  EagerImage = 'eager-image',
+}